test(plugins): split nesting-guard preflight test out of the oversized install-persistence suite (#129209)
